migrate comments
