+++ app/cmd/routes.go
@@ -110,8 +110,8 @@ func routes(r *web.Engine) *web.Engine {
 
 			private.Post("/api/v1/posts", apiv1.CreatePost())
 			private.Put("/api/v1/posts/:number", apiv1.UpdatePost())
-			private.Post("/api/posts/:number/comments", handlers.PostComment())
-			private.Post("/api/posts/:number/comments/:id", handlers.UpdateComment())
+			private.Post("/api/v1/posts/:number/comments", apiv1.PostComment())
+			private.Put("/api/v1/posts/:number/comments/:id", apiv1.UpdateComment())
 			private.Put("/api/v1/posts/:number/status", apiv1.SetResponse())
 			private.Post("/api/v1/posts/:number/support", apiv1.AddSupporter())
 			private.Delete("/api/v1/posts/:number/support", apiv1.RemoveSupporter())
